make turns on board

+++ src/main/java/com/example/demo/MovesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MovesRepository extends JpaRepository<Move, Long> {
-    List<Move> findAllByCell(Long cell);
-    List<Move> findAllByRound(Long round);
-    List<Move> findAllByCellAndRound(Long cell, long round);
+    List<Move> findAllByCell(int cell);
+    List<Move> findAllByRound(int round);
+    List<Move> findAllByCellAndRound(int cell, int round);
 }
